requirements.txt
streamlit
numpy
matplotlib
Pillow
scipy
opencv-python-headless

